requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
hubspot-api-client>=8.0.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
asyncio
pathlib
loguru>=0.7.0
pypdf>=3.0.0
python-docx>=0.8.11
faiss-cpu>=1.7.0
